Refactored findDonationsForEvent function
